+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugReactnativeModule.java
@@ -5,6 +5,7 @@ import static com.instabug.reactlibrary.utils.InstabugUtil.getMethod;
 import android.app.Application;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.util.Log;
 import android.view.View;
 
 import androidx.annotation.UiThread;
@@ -16,6 +17,7 @@ import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableArray;
+import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeArray;
@@ -29,6 +31,7 @@ import com.instabug.library.IssueType;
 import com.instabug.library.LogLevel;
 import com.instabug.library.ReproConfigurations;
 import com.instabug.library.core.InstabugCore;
+import com.instabug.library.featuresflags.model.IBGFeatureFlag;
 import com.instabug.library.internal.module.InstabugLocale;
 import com.instabug.library.invocation.InstabugInvocationEvent;
 import com.instabug.library.logging.InstabugLog;
@@ -36,6 +39,7 @@ import com.instabug.library.model.NetworkLog;
 import com.instabug.library.model.Report;
 import com.instabug.library.networkDiagnostics.model.NetworkDiagnosticsCallback;
 import com.instabug.library.ui.onboarding.WelcomeMessage;
+import com.instabug.library.util.InstabugSDKLogger;
 import com.instabug.reactlibrary.utils.ArrayUtil;
 import com.instabug.reactlibrary.utils.EventEmitterModule;
 import com.instabug.reactlibrary.utils.MainThreadHandler;
@@ -52,6 +56,7 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -64,7 +69,7 @@ import javax.annotation.Nullable;
  */
 public class RNInstabugReactnativeModule extends EventEmitterModule {
 
-    private static final String TAG = RNInstabugReactnativeModule.class.getSimpleName();
+    private static final String TAG = "IBG-RN-Core";
 
     private InstabugCustomTextPlaceHolder placeHolders;
     private static Report currentReport;
@@ -165,7 +170,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
     }
 
     @ReactMethod
-    public void setCodePushVersion(@Nullable String version) {
+    public void setCodePushVersion(@Nullable final String version) {
         MainThreadHandler.runOnMainThread(new Runnable() {
             @Override
             public void run() {
@@ -911,27 +916,38 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
-    /**
-     * Extracts HTTP connection properties. Request method, Headers, Date, Url and Response code
-     *
-     * @param jsonObject the JSON object containing all HTTP connection properties
-     * @throws JSONException
-     */
     @ReactMethod
-    public void networkLog(String jsonObject) throws JSONException {
-        NetworkLog networkLog = new NetworkLog();
-        String date = System.currentTimeMillis()+"";
-        networkLog.setDate(date);
-        JSONObject newJSONObject = new JSONObject(jsonObject);
-        networkLog.setUrl(newJSONObject.getString("url"));
-        networkLog.setRequest(newJSONObject.getString("requestBody"));
-        networkLog.setResponse(newJSONObject.getString("responseBody"));
-        networkLog.setMethod(newJSONObject.getString("method"));
-        networkLog.setResponseCode(newJSONObject.getInt("responseCode"));
-        networkLog.setRequestHeaders(newJSONObject.getString("requestHeaders"));
-        networkLog.setResponseHeaders(newJSONObject.getString("responseHeaders"));
-        networkLog.setTotalDuration(newJSONObject.getLong("duration"));
-        networkLog.insert();
+    public void networkLogAndroid(final String url,
+                                  final String requestBody,
+                                  final String responseBody,
+                                  final String method,
+                                  final double responseCode,
+                                  final String requestHeaders,
+                                  final String responseHeaders,
+                                  final double duration) {
+        try {
+            final String date = String.valueOf(System.currentTimeMillis());
+
+            NetworkLog networkLog = new NetworkLog();
+            networkLog.setDate(date);
+            networkLog.setUrl(url);
+            networkLog.setMethod(method);
+            networkLog.setResponseCode((int) responseCode);
+            networkLog.setTotalDuration((long) duration);
+
+            try {
+                networkLog.setRequest(requestBody);
+                networkLog.setResponse(responseBody);
+                networkLog.setRequestHeaders(requestHeaders);
+                networkLog.setResponseHeaders(responseHeaders);
+            } catch (OutOfMemoryError | Exception exception) {
+                Log.d(TAG, "Error: " + exception.getMessage() + "while trying to set network log contents (request body, response body, request headers, and response headers).");
+            }
+
+            networkLog.insert();
+        } catch (OutOfMemoryError | Exception exception) {
+            Log.d(TAG, "Error: " + exception.getMessage() + "while trying to insert a network log");
+        }
     }
 
     @UiThread
@@ -1026,6 +1042,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #addFeatureFlags(ReadableArray)}
+     */
     @ReactMethod
     public void addExperiments(final ReadableArray experiments) {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1042,6 +1061,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #removeFeatureFlags(ReadableArray)}
+     */
     @ReactMethod
     public void removeExperiments(final ReadableArray experiments) {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1058,6 +1080,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #removeAllFeatureFlags()}
+     */
     @ReactMethod
     public void clearAllExperiments() {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1065,6 +1090,59 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
             public void run() {
                 try {
                     Instabug.clearAllExperiments();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void addFeatureFlags(final ReadableMap featureFlagsMap) {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Iterator<Map.Entry<String, Object>> iterator = featureFlagsMap.getEntryIterator();
+                    ArrayList<IBGFeatureFlag> featureFlags = new ArrayList<>();
+                    while (iterator.hasNext()) {
+                        Map.Entry<String, Object> item = iterator.next();
+                        String variant = (String) item.getValue();
+                        String name = item.getKey();
+                        featureFlags.add(new IBGFeatureFlag(name, variant.isEmpty() ? null : variant));
+                    }
+                    if (!featureFlags.isEmpty()) {
+                        Instabug.addFeatureFlags(featureFlags);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void removeFeatureFlags(final ReadableArray featureFlags) {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                   ArrayList<String> stringArray = ArrayUtil.parseReadableArrayOfStrings(featureFlags);
+                    Instabug.removeFeatureFlag(stringArray);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void removeAllFeatureFlags() {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Instabug.removeAllFeatureFlags();
                 } catch (Exception e) {
                     e.printStackTrace();
                 }