+++ android/src/test/java/com/instabug/reactlibrary/util/GlobalMocks.java
@@ -7,8 +7,10 @@ import android.graphics.Bitmap;
 import android.util.Log;
 
 import com.instabug.library.networkDiagnostics.model.NetworkDiagnosticsCallback;
+import com.instabug.crash.models.IBGNonFatalException;
 import com.instabug.reactlibrary.utils.InstabugUtil;
 
+import org.json.JSONObject;
 import org.mockito.MockedStatic;
 
 import java.lang.reflect.Method;
@@ -63,6 +65,13 @@ public class GlobalMocks {
         reflection
                 .when(() -> InstabugUtil.getMethod(Class.forName("com.instabug.library.Instabug"), "reportScreenChange", Bitmap.class, String.class))
                 .thenReturn(mReportScreenChange);
+        // reportException mock
+        Method mCrashReportException = MockReflected.class.getDeclaredMethod("reportException", JSONObject.class, boolean.class, java.util.Map.class, JSONObject.class, IBGNonFatalException.Level.class);
+        mCrashReportException.setAccessible(true);
+        reflection
+                .when(() -> InstabugUtil.getMethod(Class.forName("com.instabug.crash.CrashReporting"), "reportException", JSONObject.class,
+                        boolean.class, java.util.Map.class, JSONObject.class, IBGNonFatalException.Level.class))
+                .thenReturn(mCrashReportException);
     }
 
     public static void close() {