+++ src/modules/Instabug.ts
@@ -1,7 +1,10 @@
 import type React from 'react';
 import { Platform, findNodeHandle, processColor } from 'react-native';
 
-import type { NavigationState as NavigationStateV5 } from '@react-navigation/native';
+import type {
+  NavigationContainerRefWithCurrent,
+  NavigationState as NavigationStateV5,
+} from '@react-navigation/native';
 import type { ComponentDidAppearEvent } from 'react-native-navigation';
 import type { NavigationAction, NavigationState as NavigationStateV4 } from 'react-navigation';
 
@@ -21,6 +24,7 @@ import InstabugUtils, { stringifyIfNotString } from '../utils/InstabugUtils';
 import * as NetworkLogger from './NetworkLogger';
 import { captureUnhandledRejections } from '../utils/UnhandledRejectionTracking';
 import type { ReproConfig } from '../models/ReproConfig';
+import type { FeatureFlag } from '../models/FeatureFlag';
 
 let _currentScreen: string | null = null;
 let _lastScreen: string | null = null;
@@ -571,6 +575,19 @@ export const onStateChange = (state?: NavigationStateV5) => {
   }, 1000);
 };
 
+/**
+ * Sets a listener for screen change
+ *  @param navigationRef a refrence of a navigation container
+ *
+ */
+export const setNavigationListener = (
+  navigationRef: NavigationContainerRefWithCurrent<ReactNavigation.RootParamList>,
+) => {
+  return navigationRef.addListener('state', () => {
+    onStateChange(navigationRef.getRootState());
+  });
+};
+
 export const reportScreenChange = (screenName: string) => {
   NativeInstabug.reportScreenChange(screenName);
 };
@@ -578,6 +595,8 @@ export const reportScreenChange = (screenName: string) => {
 /**
  * Add experiments to next report.
  * @param experiments An array of experiments to add to the next report.
+ *
+ * @deprecated Please migrate to the new Feature Flags APIs: {@link addFeatureFlags}.
  */
 export const addExperiments = (experiments: string[]) => {
   NativeInstabug.addExperiments(experiments);
@@ -586,6 +605,8 @@ export const addExperiments = (experiments: string[]) => {
 /**
  * Remove experiments from next report.
  * @param experiments An array of experiments to remove from the next report.
+ *
+ * @deprecated Please migrate to the new Feature Flags APIs: {@link removeFeatureFlags}.
  */
 export const removeExperiments = (experiments: string[]) => {
   NativeInstabug.removeExperiments(experiments);
@@ -593,6 +614,8 @@ export const removeExperiments = (experiments: string[]) => {
 
 /**
  * Clear all experiments
+ *
+ * @deprecated Please migrate to the new Feature Flags APIs: {@link removeAllFeatureFlags}.
  */
 export const clearAllExperiments = () => {
   NativeInstabug.clearAllExperiments();
@@ -614,6 +637,46 @@ export const onNetworkDiagnosticsHandler = (handler?: NetworkDiagnosticsHandler)
   });
 
   NativeInstabug.setOnNetworkDiagnosticsHandler();
+};
+
+/**
+ * Add feature flags to the next report.
+ * @param featureFlags An array of feature flags to add to the next report.
+ */
+export const addFeatureFlags = (featureFlags: FeatureFlag[]) => {
+  const entries = featureFlags.map((item) => [item.name, item.variant || '']);
+  const flags = Object.fromEntries(entries);
+  NativeInstabug.addFeatureFlags(flags);
+};
+
+/**
+ * Add a feature flag to the to next report.
+ */
+export const addFeatureFlag = (featureFlag: FeatureFlag) => {
+  addFeatureFlags([featureFlag]);
+};
+
+/**
+ * Remove feature flags from the next report.
+ * @param featureFlags An array of  feature flags to remove from the next report.
+ */
+export const removeFeatureFlags = (featureFlags: string[]) => {
+  NativeInstabug.removeFeatureFlags(featureFlags);
+};
+
+/**
+ * Remove a feature flag from the next report.
+ * @param name the name of the feature flag to remove from the next report.
+ */
+export const removeFeatureFlag = (name: string) => {
+  removeFeatureFlags([name]);
+};
+
+/**
+ * Clear all feature flags
+ */
+export const removeAllFeatureFlags = () => {
+  NativeInstabug.removeAllFeatureFlags();
 };
 
 /**